server/requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
psycopg2==2.9.3
python-dotenv==0.20.0
jwt==1.3.1
bcrypt==3.2.2